+++ packages/server/src/webapp.js
@@ -5,12 +5,15 @@ const log = require('@home-gallery/logger')('server.webapp')
 
 const { useIf, isIndex } = require('./utils')
 
-const injectStateMiddleware = (indexFile, getState, injectRemoteConsole) => {
+const injectStateMiddleware = (indexFile, getState, {basePath, injectRemoteConsole}) => {
   return (req, res) => {
     fs.readFile(indexFile, 'utf8', (err, data) => {
       if (err) {
         log.error(err, `Could not read index file ${indexFile}`)
         return res.status(404).json({error: `${err}`});
+      }
+      if (basePath && basePath != '/') {
+        data = data.replace('<base href="/">', `<base href="${basePath}">`)
       }
       if (injectRemoteConsole) {
         data = data.replace('</head>', '<script src="/remote-console.js"></script></head>')
@@ -25,9 +28,9 @@ const injectStateMiddleware = (indexFile, getState, injectRemoteConsole) => {
   }
 }
 
-const app = (webappDir, getState, injectRemoteConsole) => {
+const app = (webappDir, getState, options) => {
   const indexFile = path.resolve(webappDir, 'index.html')
-  const injectState = injectStateMiddleware(indexFile, getState, injectRemoteConsole)
+  const injectState = injectStateMiddleware(indexFile, getState, options)
   return [
     useIf(injectState, isIndex),
     (_, res) => res.sendFile(indexFile)